feat: processing items in the autocrafting monitor now show the machine

+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/task/InternalTaskPattern.java
@@ -80,8 +80,8 @@ class InternalTaskPattern extends AbstractTaskPattern {
 
     @Override
     double getPercentageCompleted() {
-        final double iterationsCompleted = originalIterationsRemaining - iterationsRemaining;
-        return iterationsCompleted / originalIterationsRemaining;
+        final long iterationsCompleted = originalIterationsRemaining - iterationsRemaining;
+        return iterationsCompleted / (double) originalIterationsRemaining;
     }
 
     protected PatternStepResult useIteration() {
